move input to window
