--- Engine/Core/Input/CMakeLists.txt
@@ -1,4 +1,0 @@
-# created on 2021/7/26
-# author @zoloypzuo
-add_library(Input Input.cpp)
-target_link_libraries(Input PUBLIC Window)
